+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Fix image card responsiveness (PR #1055)
 * Make contextual title accept and present lang parameter (PR #1056)
+* Add summary-list component based on GOV.UK Frontend (PR #1061)
 
 ## 18.1.2